+++ app/Models/Library/LibraryBook.php
@@ -2,12 +2,11 @@
 
 namespace App\Models\Library;
 
-use App;
+use App\Models\Traits\HasLanguageCodeField;
 use Iatstuti\Database\Support\NullableFields;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Log;
-use Languages;
 use Nicebooks\Isbn\Exception\InvalidIsbnException;
 use Nicebooks\Isbn\Exception\IsbnNotConvertibleException;
 use Nicebooks\Isbn\Isbn;
@@ -15,6 +14,7 @@ use Nicebooks\Isbn\Isbn;
 class LibraryBook extends Model
 {
     use NullableFields;
+    use HasLanguageCodeField;
 
     protected $nullable = [
         'author',
@@ -69,26 +69,6 @@ class LibraryBook extends Model
         if ($this->isbn13 != null) {
             $isbn = Isbn::of($this->isbn13);
             return $isbn->format();
-        }
-        return null;
-    }
-
-    public function setLanguageAttribute($value)
-    {
-        if ($value !== null) {
-            $this->language_code = Languages::lookup(null, App::getLocale())
-                ->map(fn ($l) => strtolower($l))
-                ->flip()
-                ->get(strtolower($value));
-        } else {
-            $this->language_code = null;
-        }
-    }
-
-    public function getLanguageAttribute()
-    {
-        if ($this->language_code !== null) {
-            return Languages::lookup([$this->language_code], App::getLocale())->first();
         }
         return null;
     }